+++ doc/source/controller-api.rst
@@ -1,0 +1,5 @@
+Juju Controller
+---------------
+
+.. automodule:: zaza.controller
+   :members:
